add support for local directory as a backend (#14)
This adds support for binary caches that are stored in a local
directory, similar to how nix supports local directories with `nix
copy`. They can be used by setting the binary cache URL to
`file:///path/to/cache`. This is especially useful when hosting a binary
cache and running nar-serve on the same host, because this eliminates
the need to transfer the nar over HTTP and possibly encrypting and
decrypting it when using TLS.